fix: Updated DanFever to use avoid using overwrite=True and ignore identical ids
