+++ results/sentence-transformers__paraphrase-multilingual-MiniLM-L12-v2/bf3bf13ab40c3157080a7ab344c831b9ad18b5eb/DanFEVER.json
@@ -1,0 +1,158 @@
+{
+  "dataset_revision": "5d01e3f6a661d48e127ab5d7e3aaa0dc8331438a",
+  "evaluation_time": 22.474260807037354,
+  "kg_co2_emissions": null,
+  "mteb_version": "1.12.48",
+  "scores": {
+    "train": [
+      {
+        "hf_subset": "default",
+        "languages": [
+          "dan-Latn"
+        ],
+        "main_score": 0.36872,
+        "map_at_1": 0.27985,
+        "map_at_10": 0.34371,
+        "map_at_100": 0.34545,
+        "map_at_1000": 0.34547,
+        "map_at_20": 0.34495,
+        "map_at_3": 0.33582,
+        "map_at_5": 0.34082,
+        "mrr_at_1": 0.28071551859406874,
+        "mrr_at_10": 0.3440133723869788,
+        "mrr_at_100": 0.3457450662961983,
+        "mrr_at_1000": 0.3457687882868506,
+        "mrr_at_20": 0.3452497167845053,
+        "mrr_at_3": 0.33610544484544064,
+        "mrr_at_5": 0.3411266279617126,
+        "nauc_map_at_1000_diff1": 0.3327610699866543,
+        "nauc_map_at_1000_max": 0.3239225724180551,
+        "nauc_map_at_1000_std": -0.06253399789275657,
+        "nauc_map_at_100_diff1": 0.33279998037023767,
+        "nauc_map_at_100_max": 0.32396468657000926,
+        "nauc_map_at_100_std": -0.06254838555901512,
+        "nauc_map_at_10_diff1": 0.335084487438961,
+        "nauc_map_at_10_max": 0.3257336405379233,
+        "nauc_map_at_10_std": -0.0656477521360409,
+        "nauc_map_at_1_diff1": 0.41312110825244575,
+        "nauc_map_at_1_max": 0.3002646494778475,
+        "nauc_map_at_1_std": -0.09424474699060845,
+        "nauc_map_at_20_diff1": 0.33346101093379965,
+        "nauc_map_at_20_max": 0.32448554995110424,
+        "nauc_map_at_20_std": -0.06334549591139443,
+        "nauc_map_at_3_diff1": 0.3438539741651702,
+        "nauc_map_at_3_max": 0.33011090530214043,
+        "nauc_map_at_3_std": -0.07594226625132537,
+        "nauc_map_at_5_diff1": 0.3389509206701722,
+        "nauc_map_at_5_max": 0.3283128637622002,
+        "nauc_map_at_5_std": -0.06988321674323271,
+        "nauc_mrr_at_1000_diff1": 0.3313191404203475,
+        "nauc_mrr_at_1000_max": 0.32234540005831575,
+        "nauc_mrr_at_1000_std": -0.0632445056288108,
+        "nauc_mrr_at_100_diff1": 0.3313580402693744,
+        "nauc_mrr_at_100_max": 0.32238750294718194,
+        "nauc_mrr_at_100_std": -0.06325887894546574,
+        "nauc_mrr_at_10_diff1": 0.33363460885243607,
+        "nauc_mrr_at_10_max": 0.32416185528874353,
+        "nauc_mrr_at_10_std": -0.06634033456081433,
+        "nauc_mrr_at_1_diff1": 0.409972515037506,
+        "nauc_mrr_at_1_max": 0.2975875929244888,
+        "nauc_mrr_at_1_std": -0.09548197961386379,
+        "nauc_mrr_at_20_diff1": 0.33201144055327625,
+        "nauc_mrr_at_20_max": 0.3229138013996136,
+        "nauc_mrr_at_20_std": -0.06403971938252714,
+        "nauc_mrr_at_3_diff1": 0.34244436443208276,
+        "nauc_mrr_at_3_max": 0.32853703619185154,
+        "nauc_mrr_at_3_std": -0.07669527504270128,
+        "nauc_mrr_at_5_diff1": 0.3374949161291311,
+        "nauc_mrr_at_5_max": 0.32677136193950873,
+        "nauc_mrr_at_5_std": -0.07054328878311537,
+        "nauc_ndcg_at_1000_diff1": 0.29420001016518826,
+        "nauc_ndcg_at_1000_max": 0.31853831634082874,
+        "nauc_ndcg_at_1000_std": -0.03464030338992844,
+        "nauc_ndcg_at_100_diff1": 0.295210063914423,
+        "nauc_ndcg_at_100_max": 0.3196359280088633,
+        "nauc_ndcg_at_100_std": -0.03506550091145087,
+        "nauc_ndcg_at_10_diff1": 0.30582881094322084,
+        "nauc_ndcg_at_10_max": 0.32806699623570273,
+        "nauc_ndcg_at_10_std": -0.0489109682826596,
+        "nauc_ndcg_at_1_diff1": 0.41286238999542724,
+        "nauc_ndcg_at_1_max": 0.3001960984742734,
+        "nauc_ndcg_at_1_std": -0.09382379722492389,
+        "nauc_ndcg_at_20_diff1": 0.2999005832464326,
+        "nauc_ndcg_at_20_max": 0.32346003027707343,
+        "nauc_ndcg_at_20_std": -0.04059697807077387,
+        "nauc_ndcg_at_3_diff1": 0.3240528104625311,
+        "nauc_ndcg_at_3_max": 0.3375735301863195,
+        "nauc_ndcg_at_3_std": -0.07005074099346449,
+        "nauc_ndcg_at_5_diff1": 0.31505882879802266,
+        "nauc_ndcg_at_5_max": 0.3341912183101454,
+        "nauc_ndcg_at_5_std": -0.05919908728268741,
+        "nauc_precision_at_1000_diff1": 0.16112821454495438,
+        "nauc_precision_at_1000_max": 0.28416003666293316,
+        "nauc_precision_at_1000_std": 0.0713032055560556,
+        "nauc_precision_at_100_diff1": 0.1691554329597034,
+        "nauc_precision_at_100_max": 0.29269350619919454,
+        "nauc_precision_at_100_std": 0.06763741036491064,
+        "nauc_precision_at_10_diff1": 0.21836401443730669,
+        "nauc_precision_at_10_max": 0.3313743640727248,
+        "nauc_precision_at_10_std": 0.005431740353825924,
+        "nauc_precision_at_1_diff1": 0.41286238999542724,
+        "nauc_precision_at_1_max": 0.3001960984742734,
+        "nauc_precision_at_1_std": -0.09382379722492389,
+        "nauc_precision_at_20_diff1": 0.19480458396711972,
+        "nauc_precision_at_20_max": 0.3133258175807151,
+        "nauc_precision_at_20_std": 0.037949698953571094,
+        "nauc_precision_at_3_diff1": 0.26899350609769396,
+        "nauc_precision_at_3_max": 0.3574248366558109,
+        "nauc_precision_at_3_std": -0.05280454294943137,
+        "nauc_precision_at_5_diff1": 0.24683587340822324,
+        "nauc_precision_at_5_max": 0.3492133194465745,
+        "nauc_precision_at_5_std": -0.02679521598194119,
+        "nauc_recall_at_1000_diff1": 0.16243741884039248,
+        "nauc_recall_at_1000_max": 0.28384927624730644,
+        "nauc_recall_at_1000_std": 0.06914276959096431,
+        "nauc_recall_at_100_diff1": 0.17048098930808925,
+        "nauc_recall_at_100_max": 0.29240310591505997,
+        "nauc_recall_at_100_std": 0.06547223896948845,
+        "nauc_recall_at_10_diff1": 0.21942216170055293,
+        "nauc_recall_at_10_max": 0.3310377974846271,
+        "nauc_recall_at_10_std": 0.0035287664033749706,
+        "nauc_recall_at_1_diff1": 0.41312110825244575,
+        "nauc_recall_at_1_max": 0.3002646494778475,
+        "nauc_recall_at_1_std": -0.09424474699060845,
+        "nauc_recall_at_20_diff1": 0.19583367708906083,
+        "nauc_recall_at_20_max": 0.3129598924604191,
+        "nauc_recall_at_20_std": 0.03608408842679486,
+        "nauc_recall_at_3_diff1": 0.2695724111226308,
+        "nauc_recall_at_3_max": 0.3577472208844836,
+        "nauc_recall_at_3_std": -0.05359753106132794,
+        "nauc_recall_at_5_diff1": 0.247585192478047,
+        "nauc_recall_at_5_max": 0.3493996538220437,
+        "nauc_recall_at_5_std": -0.027953034312053603,
+        "ndcg_at_1": 0.27993,
+        "ndcg_at_10": 0.36872,
+        "ndcg_at_100": 0.37669,
+        "ndcg_at_1000": 0.37733,
+        "ndcg_at_20": 0.37323,
+        "ndcg_at_3": 0.35272,
+        "ndcg_at_5": 0.36171,
+        "precision_at_1": 0.27993,
+        "precision_at_10": 0.0445,
+        "precision_at_100": 0.00481,
+        "precision_at_1000": 0.00049,
+        "precision_at_20": 0.02314,
+        "precision_at_3": 0.13379,
+        "precision_at_5": 0.08464,
+        "recall_at_1": 0.27985,
+        "recall_at_10": 0.44461,
+        "recall_at_100": 0.48078,
+        "recall_at_1000": 0.4858,
+        "recall_at_20": 0.4625,
+        "recall_at_3": 0.40122,
+        "recall_at_5": 0.42296
+      }
+    ]
+  },
+  "task_name": "DanFEVER"
+}
