+++ results/sentence-transformers__paraphrase-multilingual-MiniLM-L12-v2/bf3bf13ab40c3157080a7ab344c831b9ad18b5eb/model_meta.json
@@ -1,1 +1,1 @@
-{"name": "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2", "revision": "bf3bf13ab40c3157080a7ab344c831b9ad18b5eb", "release_date": null, "languages": [], "n_parameters": null, "memory_usage": null, "max_tokens": null, "embed_dim": null, "license": null, "open_source": null, "similarity_fn_name": null, "framework": ["Sentence Transformers"], "loader": null}
+{"name": "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2", "revision": "bf3bf13ab40c3157080a7ab344c831b9ad18b5eb", "release_date": "2019-11-01", "languages": ["ara_Arab", "bul_Cyrl", "cat_Latn", "ces_Latn", "dan_Latn", "deu_Latn", "ell_Grek", "eng_Latn", "spa_Latn", "est_Latn", "fas_Arab", "fin_Latn", "fra_Latn", "fra_Latn", "glg_Latn", "guj_Gujr", "heb_Hebr", "hin_Deva", "hrv_Latn", "hun_Latn", "hye_Armn", "ind_Latn", "ita_Latn", "jpn_Jpan", "kat_Geor", "kor_Hang", "kur_Arab", "lit_Latn", "lav_Latn", "mkd_Cyrl", "mon_Cyrl", "mar_Deva", "msa_Latn", "mya_Mymr", "nob_Latn", "nld_Latn", "pol_Latn", "por_Latn", "por_Latn", "ron_Latn", "rus_Cyrl", "slk_Latn", "slv_Latn", "sqi_Latn", "srp_Cyrl", "swe_Latn", "tha_Thai", "tur_Latn", "ukr_Cyrl", "urd_Arab", "vie_Latn", "zho_Hans", "zho_Hant"], "n_parameters": null, "memory_usage": null, "max_tokens": null, "embed_dim": null, "license": null, "open_source": true, "similarity_fn_name": null, "framework": [], "loader": null}
